input: Always serialize modifiers
